rename pageheader to sidebar
