+++ web/components/Sidebar/Sidebar.module.css
@@ -1,4 +1,4 @@
-.pageHeader {
+.sidebar {
     background-color: var(--polar0);
     box-shadow: 0 0 15px 0px var(--polar3);
     max-width: 500px;
