+++ web/components/index.tsx
@@ -2,5 +2,5 @@ export {default as Board} from './Board/Board'
 export {default as Card} from './Card/Card'
 export {default as CardForm} from './CardForm/CardForm'
 export {default as Dropzone} from './Dropzone/Dropzone'
-export {default as PageHeader} from './PageHeader/PageHeader'
+export {default as Sidebar} from './Sidebar/Sidebar'
 export {default as SortableList} from './SortableList/SortableList'
